+++ extensions/anybox/src/saveNote.tsx
@@ -1,11 +1,10 @@
 import { ActionPanel, Action, Form, Clipboard, Icon, useNavigation, getPreferenceValues, open } from "@raycast/api";
-import { postAndCloseMainWindow, fetchCollections } from "./utilities/fetch";
+import { postAndCloseMainWindow, fetchTags } from "./utilities/fetch";
 import { useState, useEffect } from "react";
 
-interface CollectionProp {
+interface TagProp {
   id: string;
   name: string;
-  heading?: string;
 }
 
 interface Preferences {
@@ -22,7 +21,7 @@ function NoteForm() {
   const preferences = getPreferenceValues<Preferences>();
   const [note, setNote] = useState<string>();
   const [comment, setComment] = useState<string>();
-  const [collections, setCollections] = useState<CollectionProp[]>([]);
+  const [tags, setTags] = useState<TagProp[]>([]);
   const { pop } = useNavigation();
 
   useEffect(() => {
@@ -33,19 +32,12 @@ function NoteForm() {
         }
       }
     });
-    fetchCollections().then((tags) => {
+    fetchTags().then((tags) => {
       if (Array.isArray(tags)) {
-        setCollections(tags);
+        setTags(tags);
       }
     });
   }, []);
-
-  function collectionTitle(tag: CollectionProp) {
-    if (tag.heading) {
-      return `${tag.heading} › ${tag.name}`;
-    }
-    return `${tag.name}`;
-  }
 
   return (
     <Form
@@ -60,7 +52,7 @@ function NoteForm() {
                 const data = {
                   note,
                   comment: values.comment,
-                  collections: values.collections,
+                  tags: values.tags,
                   starred: !!values.starred,
                 };
                 await postAndCloseMainWindow("save", data);
@@ -77,7 +69,7 @@ function NoteForm() {
                 const data = {
                   note,
                   comment: values.comment,
-                  collections: values.collections,
+                  tags: values.tags,
                   starred: !!values.starred,
                 };
                 const result = (await postAndCloseMainWindow("save", data)) as SaveNoteResponse;
@@ -93,9 +85,9 @@ function NoteForm() {
     >
       <Form.TextArea title="Note" id="note" value={note} onChange={setNote} />
       <Form.TextArea title="Comment" id="comment" value={comment} onChange={setComment} />
-      <Form.TagPicker id="collections" title="Collections" defaultValue={[]}>
-        {collections.map((val) => {
-          return <Form.TagPicker.Item value={val.id} title={collectionTitle(val)} key={val.id} />;
+      <Form.TagPicker id="tags" title="Tags" defaultValue={[]}>
+        {tags.map((val) => {
+          return <Form.TagPicker.Item value={val.id} title={val.name} key={val.id} />;
         })}
       </Form.TagPicker>
       <Form.Checkbox id="starred" label="Starred" defaultValue={false} />